refactor(backend): extract auth config module
